Refactor: Added class button, and changed id names.
The X and O buttons (in the message where the user chooses what to play as) are given class choice.
Their ids are changed from x and o to X and O.
The class is added to detect a click on either button.
Ids are capitalized in order to use the id name to directly decide what to print on the grid without adding if statement.