Bug: Fixed a bug where CPU won't play unless the player makes a valid move.
The bug was that when the player clicked an already occupied sqaure or clicked on a location in the canvas that is not a sqaure, the CPU would make a move, which could lead to CPU playing multiple time in a row.
This was fixed by adding a statement that checks for the clicked location.
The clicked location should still be available in locations array.
If the clicked location is not in locations, the CPU will wait for a valid entry from the user.