Feat: CPU stops playing when the player wins. checkWin returns the winning array.
When the player wins, the CPU stops playing.
The player can still play if CPU wins, but this will be resolved when game restarts (not yet inplemented).
The winniny array contains the points which make up a whole row, colomn or diagonal.
This winning array is needed to draw a line on the winning sequence.