Feat: CPU AI implemented.
CPU AI now implemented, which makes harder for the user to win (Player can still win).
CPU thinks one step ahead, it can decide on a play to win, and it can anticipate a player's win and intercept it.
If neither the CPU or player has a chance to win in the next move, the CPU plays randomly.
The algorithm was done in cpuPlay function.